+++ app/src/lib/app-state.ts
@@ -26,6 +26,7 @@ import { CloneRepositoryTab } from '../models/clone-repository-tab'
 import { BranchesTab } from '../models/branches-tab'
 import { PullRequest } from '../models/pull-request'
 import { IAuthor } from '../models/author'
+import { ComparisonCache } from './comparison-cache'
 
 export { ICommitMessage }
 
@@ -662,37 +663,6 @@ export interface ICompareState {
    * A local cache of ahead/behind computations to compare other refs to the current branch
    */
   readonly aheadBehindCache: ComparisonCache
-}
-
-export class ComparisonCache {
-  backingStore = new Map<string, IAheadBehind>()
-
-  static getKey(from: string, to: string) {
-    return `${from}...${to}`
-  }
-
-  public set(from: string, to: string, value: IAheadBehind) {
-    const key = ComparisonCache.getKey(from, to)
-    this.backingStore.set(key, value)
-  }
-
-  public get(from: string, to: string) {
-    const key = ComparisonCache.getKey(from, to)
-    return this.backingStore.get(key)
-  }
-
-  public has(from: string, to: string) {
-    const key = ComparisonCache.getKey(from, to)
-    return this.backingStore.has(key)
-  }
-
-  public get size() {
-    return this.backingStore.size
-  }
-
-  public clear() {
-    this.backingStore.clear()
-  }
 }
 
 export enum CompareActionKind {
